pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-security-capability-monitor"
version = "0.1.0"
description = "Real-time tracker for AI offensive security capabilities"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Phoenix Assistant" }]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "playwright>=1.40.0",
    "feedparser>=6.0.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "pytest-httpx>=0.30", "ruff>=0.3.0"]

[project.scripts]
aiscm = "aiscm.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["aiscm"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
